UserDAO 예상 구현
