+++ src/main/java/dao/OptionDAO.java
@@ -1,11 +1,11 @@
-package persistence.dao;
+package dao;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
-import persistence.dto.MenuRegistDTO;
+import persistence.dto.OptionDTO;
 
-public class MenuRegistDAO {
+public class OptionDAO {
     private SqlSessionFactory sqlSessionFactory = null;
-    public MenuRegistDAO(SqlSessionFactory sqlSessionFactory){
+    public OptionDAO(SqlSessionFactory sqlSessionFactory){
         this.sqlSessionFactory = sqlSessionFactory;
     }
 }
