+++ src/main/java/dao/UserDAO.java
@@ -1,4 +1,4 @@
-package persistence.dao;
+package dao;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 import persistence.dto.UserDTO;
